Handle event animations
